refactor: fix comments, move code from bulk-delete file to delete file
