+++ resources.go
@@ -103,12 +103,12 @@ func (a *App) createResources() error {
 			Put:  "/:id",
 			Args: fs.Args{"id": createArg(fs.TypeUint64, "The content ID")},
 		})).
+		Add(fs.NewResource("bulk-delete", contentService.BulkDelete, &fs.Meta{
+			Delete: "/delete",
+		})).
 		Add(fs.NewResource("delete", contentService.Delete, &fs.Meta{
 			Delete: "/:id",
 			Args:   fs.Args{"id": createArg(fs.TypeUint64, "The content ID")},
-		})).
-		Add(fs.NewResource("bulk-delete", contentService.BulkDelete, &fs.Meta{
-			Post: "/bulk-delete",
 		}))
 
 	a.api.Group("role").
