+++ services/content/delete.go
@@ -42,3 +42,47 @@ func (cs *ContentService) Delete(c fs.Context, _ any) (any, error) {
 
 	return schema.NewEntity(uint64(id)), nil
 }
+
+func (cs *ContentService) BulkDelete(c fs.Context, _ any) (int, error) {
+	schemaName := c.Arg("schema")
+	model, err := cs.DB().Model(schemaName)
+	if err != nil {
+		return 0, errors.BadRequest(err.Error())
+	}
+
+	predicates, err := db.CreatePredicatesFromFilterObject(
+		cs.DB().SchemaBuilder(),
+		model.Schema(),
+		c.Arg("filter"),
+	)
+	if err != nil {
+		return 0, errors.BadRequest(err.Error())
+	}
+
+	records, err := model.Query(predicates...).
+		Get(c.Context())
+
+	if err != nil {
+		return 0, errors.BadRequest(err.Error())
+	}
+
+	if len(records) == 0 {
+		return 0, nil
+	}
+
+	var ids []any
+	for _, record := range records {
+		recordID := record.ID()
+		if err := isDeletable(int(recordID), schemaName); err != nil {
+			return 0, errors.InternalServerError("Cannot delete root user.")
+		}
+		ids = append(ids, int(recordID))
+	}
+
+	recordDelete, err := model.Mutation().Where(db.In("id", ids)).Delete(c.Context())
+	if err != nil {
+		return 0, errors.InternalServerError(err.Error())
+	}
+
+	return recordDelete, nil
+}
